// web-client/src/app/@core/models/history-change-event.enum.ts
export enum HistoryChangeEventEnum {
    HIDE_POST = 'Удаление должности',
    UPDATE_POST = 'Обновление должности',
    HIDE_DEPARTMENT = 'Удаление подразделения',
    UPDATE_DEPARTMENT = 'Обновление подразделения',
    HIDE_EMPLOYEE = 'Удаление сотрудника',
    UPDATE_EMPLOYEE = 'Обновление сотрудника',
    ADD_DEPARTMENT = 'Добавление подразделения',
    ADD_POST = 'Добавление дожлности',
    ADD_EMPLOYEE = 'Добавление сотрудника'
}